Add abstract classes for value-to-null mutators/tests
