+++ .travis.yml
@@ -29,7 +29,7 @@ script:
   - vendor/bin/phpunit --coverage-clover=clover.xml
   - /home/travis/.composer/vendor/bin/php-cs-fixer fix --config=.php_cs.dist -v --dry-run --stop-on-violation --using-cache=no --path-mode=intersection --allow-risky=yes
   - /home/travis/.composer/vendor/bin/phpstan analyse src tests --level=1 --no-interaction --no-progress
-  - if [[ $INFECTION == 1 ]]; then bin/infection --threads=4 --min-msi=46 --min-covered-msi=88; fi
+  - if [[ $INFECTION == 1 ]]; then bin/infection --threads=4 --min-msi=47 --min-covered-msi=88; fi
 
 after_success:
   - bash <(curl -s https://codecov.io/bash)
