+++ src/Mutator/ReturnValue/NewObject.php
@@ -11,7 +11,7 @@ namespace Infection\Mutator\ReturnValue;
 use Infection\Mutator\FunctionBodyMutator;
 use PhpParser\Node;
 
-class NewObject extends FunctionBodyMutator
+class NewObject extends AbstractValueToNullReturnValue
 {
     /**
      * Replace "return new Something(anything);" with "new Something(anything); return null;"
@@ -44,6 +44,6 @@ class NewObject extends FunctionBodyMutator
             return false;
         }
 
-        return true;
+        return $this->isNullReturnValueAllowed($node);
     }
 }
