+++ tests/Mutator/ReturnValue/AbstractValueToNullReturnValueTest.php
@@ -4,19 +4,13 @@
  *
  * License: https://opensource.org/licenses/BSD-3-Clause New BSD License
  */
-
 declare(strict_types=1);
 
-namespace Infection\Tests\Mutator\ReturnValue;
+namespace Mutator\ReturnValue;
 
-use Infection\Mutator\Mutator;
-use Infection\Mutator\ReturnValue\FloatNegation;
-use Infection\Mutator\ReturnValue\IntegerNegation;
-use Infection\Mutator\ReturnValue\NewObject;
 use Infection\Tests\Mutator\AbstractMutator;
 
-
-class NewObjectTest extends AbstractMutator
+abstract class AbstractValueToNullReturnValueTest extends AbstractMutator
 {
     public function test_not_mutates_with_value_return_true()
     {
@@ -41,7 +35,6 @@ CODE;
         $this->assertSame($expectedMutatedCode, $mutatedCode);
     }
 
-
     public function test_not_mutates_with_value_return_true_for_new_objects()
     {
         $code = <<<'CODE'
@@ -63,33 +56,5 @@ function test()
 CODE;
 
         $this->assertSame($expectedMutatedCode, $mutatedCode);
-    }
-
-    public function test_mutates_instantiation_of_new_object_with_params()
-    {
-        $code = <<<'CODE'
-<?php
-function test()
-{
-    return new Foo('now');
-}
-CODE;
-        $mutatedCode = $this->mutate($code);
-
-        $expectedMutatedCode = <<<'CODE'
-<?php
-
-function test()
-{
-    return new Foo('now');
-}
-CODE;
-
-        $this->assertSame($expectedMutatedCode, $mutatedCode);
-    }
-
-    protected function getMutator() : Mutator
-    {
-        return new NewObject();
     }
 }
